--- holiday_cal/credentials.py
@@ -1,16 +1,0 @@
-""" apis and urls"""
-import os
-api_key = ['YOUR API KEY HERE']
-assert api_key is not None
-
-
-# url
-url_countries = 'https://calendarific.com/api/v2/countries'
-url_holiday = f'https://calendarific.com/api/v2/holidays'
-
-
-
-
-
-
-
