--- holiday_cal/scratch.json
@@ -1,6 +1,0 @@
-{
-  "holiday_name": "Day off for New Year's Day",
-  "description": "New Year's Day is the first day of the Gregorian calendar, which is widely used in many countries such as the USA.",
-  "country": "United States",
-  "date": "2021-12-31"
-}
